requirements.txt
boto3
pandas
pyspark
mysql-connector-python
cryptography
findspark 